Separate specification from code generation (fix #376)
Move all code generation into module code_gen, which replaces run_fn.
