+++ src/Makefile.am
@@ -17,7 +17,7 @@ DIST_SRCS = \
 	stack.py \
 	spec.py \
 	code_util.py \
-	run_fn.py \
+	code_gen.py \
 	gen-opcodes \
 	gen-instructions \
 	gen-main \
@@ -69,7 +69,7 @@ include/mit/opcodes.h: gen-opcodes spec.py action.py
 	$(MKDIR_P) include/mit
 	$(PYTHON_WITH_PATH) $(srcdir)/gen-opcodes > include/mit/opcodes.h || ( rm -f include/mit/opcodes.h; exit 1 )
 
-instructions.c: gen-instructions spec.py action.py stack.py code_util.py run_fn.py
+instructions.c: gen-instructions spec.py action.py stack.py code_util.py code_gen.py
 	$(PYTHON_WITH_PATH) $(srcdir)/gen-instructions > instructions.c || ( rm -f instructions.c; exit 1 )
 
 main.c: gen-main
