+++ src/specializer/specializer.am
@@ -84,7 +84,7 @@ specialize-once: mit@PACKAGE_SUFFIX@$(EXEEXT)
 %D%/labels.json: code_util.py action.py spec.py %D%/profile.json %D%/simulate-jit %D%/path.py %D%/profile.py
 	$(PYTHON_WITH_PATH) $(srcdir)/%D%/simulate-jit $(srcdir)/%D%/profile.json $@
 
-%D%/specializer.c: code_util.py action.py stack.py spec.py run_fn.py %D%/specializer.am %D%/gen-specializer %D%/specializer.py %D%/specializer_spec.py %D%/labels.json
+%D%/specializer.c: code_util.py action.py stack.py spec.py code_gen.py %D%/specializer.am %D%/gen-specializer %D%/specializer.py %D%/specializer_spec.py %D%/labels.json
 	$(PYTHON_WITH_PATH) $(srcdir)/%D%/gen-specializer %D%/labels.json > %D%/specializer.c || ( rm -f %D%/specializer.c; exit 1 )
 
 %D%/specializer.lo: include/mit/mit.h
