+++ src/stack.py
@@ -1,7 +1,7 @@
 '''
 Model and generate code for the VM stack.
 
-Copyright (c) 2009-2019 Mit authors
+Copyright (c) 2009-2020 Mit authors
 
 The package is distributed under the MIT/X11 License.
 
@@ -12,7 +12,7 @@ RISK.
 import re
 from functools import total_ordering
 
-from code_util import Code, unrestrict, disable_warnings
+from code_util import unrestrict
 
 
 # Enough for the core
@@ -33,69 +33,6 @@ def type_words(type):
         import sys
         print(f'type {type} not found; type_words: "{type_wordses}"', file=sys.stderr)
     return ret
-
-def load_stack(name, depth=0, type='mit_word_t'):
-    '''
-    Generate C code to load the variable `name` of type `type` occupying
-    stack slots `depth`, `depth+1`, ... . Does not check the stack.
-
-    Returns a Code.
-    '''
-    code = Code()
-    code.append(
-        f'mit_max_stack_item_t temp = (mit_uword_t)(*mit_stack_pos(stack, stack_depth, {depth}));'
-    )
-    for i in range(1, type_words(type)):
-        code.append('temp <<= MIT_WORD_BIT;')
-        code.append(
-            f'temp |= (mit_uword_t)(*mit_stack_pos(stack, stack_depth, {depth + i}));'
-        )
-    code.extend(disable_warnings(
-        ['-Wint-to-pointer-cast'],
-        Code(f'{name} = ({type})temp;'),
-    ))
-    return Code('{', code, '}')
-
-def store_stack(value, depth=0, type='mit_word_t'):
-    '''
-    Generate C code to store the value `value` of type `type` occupying
-    stack slots `depth`, `depth+1`, ... . Does not check the stack.
-
-    Returns a Code.
-    '''
-    code = Code()
-    code.extend(disable_warnings(
-        ['-Wpointer-to-int-cast', '-Wbad-function-cast'],
-        Code(f'mit_max_stack_item_t temp = (mit_max_stack_item_t){value};'),
-    ))
-    for i in reversed(range(1, type_words(type))):
-        code.append(
-            f'*mit_stack_pos(stack, stack_depth, {depth + i}) = (mit_uword_t)(temp & MIT_UWORD_MAX);'
-        )
-        code.append('temp >>= MIT_WORD_BIT;')
-    code.append(
-        '*mit_stack_pos(stack, stack_depth, {}) = (mit_uword_t)({});'
-        .format(
-            depth,
-            'temp & MIT_UWORD_MAX' if type_words(type) > 1 else 'temp',
-        )
-    )
-    return Code('{', code, '}')
-
-def pop_stack(name, type='mit_word_t'):
-    code = Code()
-    code.extend(check_underflow(Size(type_words(type))))
-    code.extend(load_stack(name, type=type))
-    code.append(f'stack_depth -= {type_words(type)};')
-    return code
-
-def push_stack(value, type='mit_word_t'):
-    code = Code()
-    code.extend(check_overflow(Size(0), Size(type_words(type))))
-    code.extend(store_stack(value, depth=-type_words(type), type=type))
-    code.append(f'stack_depth += {type_words(type)};')
-    return code
-
 
 @total_ordering
 class Size:
@@ -217,18 +154,6 @@ class StackItem:
     def __hash__(self):
         return hash((self.name, self.type, self.size))
 
-    def load(self):
-        '''
-        Returns a Code to load `self` from the stack to its C variable.
-        '''
-        return load_stack(self.name, self.depth, self.type)
-
-    def store(self):
-        '''
-        Returns a Code to store `self` to the stack from its C variable.
-        '''
-        return store_stack(self.name, self.depth, self.type)
-
 
 class StackPicture:
     '''
@@ -317,84 +242,3 @@ class StackEffect:
          - results - list acceptable to `StackPicture.of()`.
         '''
         return StackEffect(StackPicture.of(args), StackPicture.of(results))
-
-    def declare_vars(self):
-        '''
-        Returns a Code to declare C variables for arguments and results other
-        than 'ITEMS'.
-        '''
-        return Code(*[
-            f'{item.type} {item.name};'
-            for item in self.by_name.values()
-            if item.name != 'ITEMS'
-        ])
-
-    def load_args(self):
-        '''
-        Returns a Code to read the arguments from the stack into C
-        variables, skipping 'ITEMS' and 'COUNT'.
-
-        `stack_depth` is not modified.
-        '''
-        code = Code()
-        for item in self.args.items:
-            if item.name != 'ITEMS' and item.name != 'COUNT':
-                code.extend(item.load())
-        return code
-
-    def store_results(self):
-        '''
-        Returns a Code to write the results from C variables into the stack,
-        skipping 'ITEMS'.
-
-        `stack_depth` must be modified first.
-        '''
-        code = Code()
-        for item in self.results.items:
-            if item.name != 'ITEMS':
-                code.extend(item.store())
-        return code
-
-
-def check_underflow(num_pops):
-    '''
-    Returns a Code to check that the stack contains enough items to
-    pop the specified number of items.
-     - num_pops - Size, with non-negative `count`.
-    '''
-    assert isinstance(num_pops, Size)
-    assert num_pops >= 0, num_pops
-    if num_pops == 0: return Code()
-    tests = []
-    tests.append(f'unlikely(stack_depth < (mit_uword_t)({num_pops.size}))')
-    if num_pops.count == 1:
-        tests.append(
-            f'unlikely(stack_depth - (mit_uword_t)({num_pops.size}) < (mit_uword_t)(COUNT))'
-        )
-    return Code(
-        'if ({}) {{'.format(' || '.join(tests)),
-        Code('THROW(MIT_ERROR_INVALID_STACK_READ);'),
-        '}',
-    )
-
-def check_overflow(num_pops, num_pushes):
-    '''
-    Returns a Code to check that the stack contains enough space to
-    push `num_pushes` items, given that `num_pops` items will first be
-    popped successfully.
-     - num_pops - Size.
-     - num_pushes - Size.
-    `num_pops` and `num_pushes` must both be variadic or both not.
-    '''
-    assert isinstance(num_pops, Size)
-    assert isinstance(num_pushes, Size)
-    assert num_pops >= 0
-    assert num_pushes >= 0
-    depth_change = num_pushes - num_pops
-    if depth_change <= 0: return Code()
-    # Ensure comparison will not overflow
-    assert depth_change.count == 0
-    return Code(f'''\
-        if (unlikely(stack_words - stack_depth < {depth_change}))
-            THROW(MIT_ERROR_STACK_OVERFLOW);'''
-    )
